+++ packages/lib/src/types/objectBased/typesRecord.ts
@@ -56,9 +56,9 @@ export function typesRecord<T extends AnyType>(valueType: T): RecordType<T> {
     const thisTc: TypeChecker = new TypeChecker(
       TypeCheckerBaseType.Object,
 
-      (obj, path) => {
+      (obj, path, typeCheckedValue) => {
         if (!isObject(obj)) {
-          return new TypeCheckError(path, getTypeName(thisTc), obj)
+          return new TypeCheckError(path, getTypeName(thisTc), obj, typeCheckedValue)
         }
 
         if (!valueChecker.unchecked) {
@@ -66,7 +66,7 @@ export function typesRecord<T extends AnyType>(valueType: T): RecordType<T> {
           for (let i = 0; i < keys.length; i++) {
             const k = keys[i]
             const v = obj[k]
-            const valueError = valueChecker.check(v, [...path, k])
+            const valueError = valueChecker.check(v, [...path, k], typeCheckedValue)
             if (valueError) {
               return valueError
             }